Fix build failure in CI.

+++ src/plugins/intel_gpu/src/graph/include/pass_manager.h
@@ -324,7 +324,7 @@ public:
         }
 
         auto mem_deps = node->get_memory_dependencies();
-        if (mem_deps.find(dep->get_unique_id()) != mem_deps.end()) {
+        if (mem_deps.find(static_cast<uint32_t>(dep->get_unique_id())) != mem_deps.end()) {
             return;
         }
